# Speedup `marginalmeans()`