--- a/R/get_modeldata.R
+++ b/R/get_modeldata.R
@@ -104,9 +104,9 @@ set_variable_class <- function(modeldata, model = NULL) {
             cl[col] <- "other"
         } else if (is.numeric(out[[col]])) {
             # more performant in time and memory than all(out[[col]] %in% 0:1)
-            if (min(out[[col]]) == 0 &&
-                max(out[[col]]) == 1 &&
-                length(unique(out[[col]])) == 2) {
+            if (isTRUE(min(out[[col]]) == 0 &&
+                       max(out[[col]]) == 1 &&
+                       length(unique(out[[col]])) == 2)) {
                 cl[col] <- "binary"
             } else {
                 cl[col] <- "numeric"