@vincentarelbundock 2023-08-23
Another instance, but probably much less important:

https://github.com/vincentarelbundock/marginaleffects/blob/main/R/sanitize_variables.R#L179