@etiennebacher 2023-08-23
Might be worth making a small helper `is_binary()` (FYI I won't do a PR for this though ;) )