@vincentarelbundock 2023-08-23
Hmmm, on further reflection, I think this is going to break if the vector has only one element...